Core/DBLayer: Match field type reads in core with data types in db for prepared statements

+++ src/server/game/Chat/Chat.cpp
@@ -467,7 +467,7 @@ static ChatCommand wintergraspCommandTable[] = //Línea de control chat -- Comand
                 Field* fields = result->Fetch();
                 std::string name = fields[0].GetString();
 
-                SetDataForCommandInTable(commandTableCache, name.c_str(), fields[1].GetUInt16(), fields[2].GetString(), name);
+                SetDataForCommandInTable(commandTableCache, name.c_str(), fields[1].GetUInt8(), fields[2].GetString(), name);
 
             } while (result->NextRow());
         }
